[ACC-1444] Define rabbitmq setup job as a normal job, not a hook.
Running it as a post hook does not work because some of the deployments (the ones depending on rabbitmq queues) never
become healthy without definitions present. This results in the post hooks never being run, and the install/upgrade failing

Also ensure that job name is unique when config has changed

To ensure that a new job gets created (and the old one deleted) when definitions change, add the definition hash to the job name
A job can only be created and run once, and not really modified or restarted afterwards. Only creating a new job will do for that