[ACC-1444] Do rabbitmq account password secret patching in rabbitmq setup
Trying to patch secrets that do not yet exist (during initial install or
upgrade from older version) does not end well. The pre-install hook
fails to complete because the secrets are only installed by the helm
chart afterwards.

Instead, generate the passwords during the rabbitmq setup script. That
way, they can be inserted in the secret and in rabbitmq in one stage.